requirements.txt
###############################
# --- CORE BACKEND ---
###############################
Flask==3.0.3
Flask-Cors==4.0.0
requests==2.31.0
python-dotenv==1.0.1

###############################
# --- AUTHENTICATION ---
###############################
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1

###############################
# --- DATABASE ---
###############################
mysql-connector-python==8.3.0
SQLAlchemy==2.0.34   # optional (used in some training utilities)

###############################
# --- FRONTEND (STREAMLIT) ---
###############################
streamlit==1.38.0
pandas==2.2.2
plotly==5.22.0
numpy==1.26.4

###############################
# --- NLP / ML MODEL INFERENCE ---
###############################
transformers==4.42.4
torch==2.3.0
sentencepiece==0.2.0
rouge-score==0.1.2
nltk==3.9.1
textstat==0.7.4
evaluate==0.4.3

###############################
# --- DATASETS (needed for TRAIN_MODELS scripts) ---
###############################
datasets==2.19.1
accelerate==0.30.1
huggingface-hub==0.23.2

###############################
# --- SCORING / EVALUATION ---
###############################
sacrebleu==2.4.1
scikit-learn==1.5.1  # for accuracy, metrics, vectorizers (used in evaluate_models.py)

###############################
# --- TRAINING PIPELINE ---
###############################
protobuf==5.27.1
tqdm==4.66.4

###############################
# --- OPTIONAL UTILS ---
###############################
python-multipart==0.0.9
pyyaml==6.0.1

###############################
# --- DEPLOYMENT ---
###############################
gunicorn==22.0.0
